UploadMenu: Refactor to use styled component and sx instead of makeStyles
