--- src/components/DropboxMenuItem/index.ts
@@ -1,1 +1,0 @@
-export { DropboxMenuItem } from "./DropboxMenuItem";
