+++ src/components/UploadMenu/StyledMenuItem.tsx
@@ -1,0 +1,8 @@
+import { styled } from "@mui/material/styles";
+import MenuItem, { MenuItemProps } from "@mui/material/MenuItem";
+
+type StyledMenuItemProps = MenuItemProps & { component: string };
+
+export const StyledMenuItem = styled(MenuItem)<StyledMenuItemProps>(() => ({
+  maxWidth: 320,
+}));
